Mock I2C for testing
I created a mock of I2C that allows the setting of a state for a
response. I think that the next step is to make it store any data we
want to read. For now it is very focused on reading a button.

Amos King @adkron <amos@binarynoggin.com>