Remove crash report from recovery test
The button test that makes sure that we recover from I2C errors and
still have a working button printed out crash reports.

~~~
  ERROR REPORT==== 26-Mar-2017::15:14:47 ===
  ** Generic server <0.184.0> terminating
  #...
~~~

We don't want these to show up when running the tests because they can
make people believe that we have an error when we do not. We start the
logger and then add the tag to any test that we want to suppress that
crash report output.

Amos King @adkron <amos@binarynoggin.com>