Default I2C to use elixir_ale module
This will allow clients to not have to configure their applications of
they want to use the default of elixir_ale for I2C communication. The
default has also moved into the GrovePi.I2C module so that the default
is configured in one location.

Amos King @adkron <amos@binarynoggin.com>