Update tests with new api
I left one failing test for us to discuss. The issue is that we are no
longer supervising the button ourselves so the failed i2c error is
bubbling up to our tests and crashing them. This is okay if the behavior
we want it those errors to bubble up to the user's supervisor. I see
these errors as a grove pi problem so I'm leaning that we need a
supervisor for our buttons.

Amos King @adkron <amos@binarynoggin.com>