Fix testing on OS X
Building on OS X was failing. I believe that the deps are loaded before
the env configs. The way this was is loading the default as having
elixir_ale. This won't build on OS X. For right now I thought I would
leave the dev environment building with elixir_ale so that we could have
a happy medium. I don't see us needing it in test any time soon.

Amos King @adkron <amos@binarynoggin.com>