Change button polling to use `Process.send_after`
After some thought about us being library code I thought it wouldn't be
a good idea for us to build on top of timer. The timer process can get
overloaded with messages if we are using it and the client code is using
it. Also the `send_after` will send a message even if we are still
processing the last message. Over time the last message processing will
at least creep up on our interval. In order to avoid this backing up I
decided to call `Process.send_after`. This is a single message to be
sent. In the polling function we the do a `send_after` when we have
completed the work. This means that we will start our timer again after
work is complete. Hopefully overloading won't be an issue.

Amos King @adkron <amos@binarynoggin.com>