Add time of written messages to I2C mock
The time of the written messages allows us to test things that are
supposed to be sent in certain time frames. The default behavior is
still the same as before.

~~~elixir
  message = GrovePi.I2C.get_last_write(GrovePi.Board)

  {epoch_time, message} =
    GrovePi.I2C.get_last_write(GrovePi.Board, include_time: true)
~~~

The epoch time is the time that the message was sent to the
I2C mock module.

Amos King @adkron <amos@binarynoggin.com>