Split I2C state into its own file
I wanted to do this so that I could view the i2c functionality on its
own and not get muddled up in what is happening in the state. The state
seems more like the board itself and shouldn't be part of the fake.

Amos King @adkron <amos@binarynoggin.com>