Allow specs to run async
I still need to figure out how to get the sleeps out of the start
functions in the specs. I keep getting already started exceptions when
they are not there.

This, in theory, allows for more than one board to be stacked, which
according to the grove manufacturer works, in theory. ;)

Amos King @adkron <amos@binarynoggin.com>