Use generated prefix to avoid async issues in test
This uses a prefix generated from a timestamp in the
test setup to isolate the supervision trees for each test.
This allows us to remove the sleep call in setup.
This will need to be replicated to the other tests
and refactored to pull out common code into a helper module

Todd Pickell <tapickell@gmail.com> @tapickell
Amos King <amos@binarynoggin.com> @adkron