Extract poller for polling modules
The polling modules require a few arguments. These are for setting the
default triggering mechanism and the read type. The read type is the
type of return value from the read_value callback. The read value
callback is required to be implemented. This callback is how we get the
data into the poller.

Amos King @adkron <amos@binarynoggin.com>