Allow trigger options to be passed to hysteresis trigger
The hysteresis defaults may not be right in all environments. Allowing
the user to set the high and low thresholds is important to allowing the
sound triggers to be set for their environment.

In setting the sound triggers this allows all triggers to be passed in a
set of arguments.

Amos King @adkron <amos@binarynoggin.com>