Fix already started because of module name collision times
The prefix names in tests are set based on time. This time can end up
being the same across more than one test since they are running with
`async: true`. I added the module name to the prefix because that will
make sure that across the different files we are running with different
prefixes. This works because the tests in a single file are not run
asynchronously.

Amos King @adkron <amos@binarynoggin.com>