Add poller for DHT11 sensor
Add poller support for a DHT11 sensor.  An event is created for any
change to the humidity or temperature values.  The DHT11 is read using
code copied from the DHT module (the DHT module must be refactored for
polling). The module will only work with DHT module_type 0 which
is the configuration for the DHT sensor that comes with the GrovePi+
Starter Kit.