Distribute the license and changelog
The full license text really should be included in the release package
even though it's listed in the mix.exs for completeness. Including the
changelog is convenient to have in the deps.

I also tried including the examples, but their build artifacts get
included when trying to wildcard the inclusion. Presumably anyone
building the examples will be building them from a clone of this
repository rather than from inside of another project's deps directory
anyway.