Simplify PCA9685 initialization and updates
The PCA9685 has an auto-update mode (abbreviated AI) that lets you set
multiple registers in one I2C transaction. If this is enabled, it's
possible to update LEDs in one go with a properly constructed binary
pattern. This made it possible to remove quite a few calculations and
merge the "all LEDs" register setting code.

This also includes a couple other changes:

* Added `set_pwm_on` and `set_pwm_off` to set the channels to 100% on or
  100% off. This felt more obvious than calling `set_pwm` with 0x1000.
* Updated the initialization to keep the PCA9685 off. This removes one
  of the sleeps.
* Turned off the "all call" address. This frees up I2C address 0x70 in
  case some other Grove device was going to use it.
* Renamed a few constants to make it easier to not mix up mode1 and
  mode2 bits