Updates the address to match the docs and fix writable
The address I was using was the board, and when looking back over the
docs I noticed that I had the wrong address.

Fixing the writable call in the board. I was looking back through and
realized I had left in it taking two arguments when there is definitely
only one.

Amos King @adkron <amos@binarynoggin.com>