Add support for GrovePi Oxygen sensor
Also renames atom from the ir reflective sensor for readability.

Default triggers also supported here so we can listen to events happily.

Chose to keep the analog read only and not convert values, and instead
push to application level because this facilitates data correction
from the oxygen sensor when needed.